Relocate compute client
