+++ pdchaosazure/__init__.py
@@ -4,18 +4,15 @@
 
 from typing import List
 
-from azure.mgmt.compute import ComputeManagementClient
-from chaoslib.discovery import (discover_actions, discover_probes,
-                                initialize_discovery_result)
-from chaoslib.types import (Configuration, DiscoveredActivities, Discovery,
-                            Secrets)
+from chaoslib.discovery import (discover_actions, discover_probes, initialize_discovery_result)
+from chaoslib.types import (DiscoveredActivities, Discovery)
 from logzero import logger
 
 from pdchaosazure.auth import auth
 from pdchaosazure.common.config import load_subscription_id, load_secrets
 
 __all__ = [
-    "__version__", "auth", "discover", "load_secrets", "init_client"
+    "__version__", "auth", "discover", "load_secrets"
 ]
 __version__ = '0.8.12-dev4'
 __package__ = "proofdock-chaos-azure"
@@ -30,19 +27,6 @@ def discover(discover_system: bool = True) -> Discovery:
     discovery = initialize_discovery_result(__package__, __version__, "azure")
     discovery["activities"].extend(__load_exported_activities())
     return discovery
-
-
-def init_client(experiment_secrets: Secrets, experiment_configuration: Configuration) -> ComputeManagementClient:
-    secrets = load_secrets(experiment_secrets)
-    configuration = load_subscription_id(experiment_configuration)
-
-    with auth(secrets) as authentication:
-        base_url = secrets.get('cloud').endpoints.resource_manager
-        client = ComputeManagementClient(credentials=authentication,
-                                         subscription_id=configuration.get('subscription_id'),
-                                         base_url=base_url)
-
-        return client
 
 
 ###############################################################################
