+++ pdchaosazure/machine/actions.py
@@ -5,9 +5,8 @@ from chaoslib.types import Configuration, Secrets
 from logzero import logger
 from msrestazure import azure_exceptions
 
-from pdchaosazure import init_client
 from pdchaosazure.common import cleanse, config
-from pdchaosazure.common.compute import command
+from pdchaosazure.common.compute import command, init_client
 from pdchaosazure.machine.constants import RES_TYPE_VM
 from pdchaosazure.common.resources.graph import fetch_resources
